refactor: remove requirements_dev.txt
Requirements_dev.txt contained numerous outdated, insecure and
deprecated packages. As noted by pytest-runner[1] the usage of
pytest/pytest-runner as setup_requirements or test_requirements
is deprecated and should be removed.

They advise to use bootstrap tools such as tox. Due to this,
`requirements_dev.text` is removed. Pytest is still required for
running tests and is therefore added to `tox.ini` as a dependency.
Tox is used to setup the environments and dependencies, so tox
is mentioned as a test_requires package in `setup.py`.

[1] https://pypi.org/project/pytest-runner/